requirements.txt
Django==3.2.16
django-filter==23.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
mixer==7.2.2
pep8-naming==0.13.3
pytest==7.1.3
pytest-django==4.5.2
flake8==5.0.4
flake8-docstrings==1.7.0